Move passages types to types directory
