+++ src/components/documents/DocumentPassageViewer.test.tsx
@@ -2,9 +2,8 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { render, screen, waitFor, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import type { ISearchPassage } from "@/api/passages";
 import { TopicsContext } from "@/context/TopicsContext";
-import { TFamilyDocumentPublic, TSearchResponse, TTopics } from "@/types";
+import { ISearchPassage, TFamilyDocumentPublic, TSearchResponse, TTopics } from "@/types";
 
 import { DocumentPassageViewer } from "./DocumentPassageViewer";
